feat: split otp to control and field pair
