+++ packages/core/src/useCalendar/index.ts
@@ -1,4 +1,5 @@
 export * from './useCalendar';
+export * from './useCalendarControl';
 export * from './types';
 export * from './useCalendarCell';
 export * from './useCalendarView';
