+++ packages/core/src/useOtpField/index.ts
@@ -1,3 +1,4 @@
 export * from './useOtpField';
 export * from './useOtpSlot';
+export * from './useOtpControl';
 export * from './types';
