+++ packages/core/src/useSearchField/index.ts
@@ -1,1 +1,2 @@
 export * from './useSearchField';
+export * from './useSearchControl';
